+++ SquareRunX/SquareRunX/Source/Engine/Graphics/OrthogonalCamera.cpp
@@ -3,28 +3,28 @@
 
 OrthoCamera::OrthoCamera() {}
 
-OrthoCamera::OrthoCamera(const glm::vec2& Pos, const glm::vec2& ViewSize) :
+OrthoCamera::OrthoCamera(const glm::dvec2& Pos, const glm::dvec2& ViewSize) :
 	Position(Pos), ViewSize(ViewSize)
 {}
 
 OrthoCamera::~OrthoCamera() {}
 
-void OrthoCamera::SetPosition(const glm::vec2& Position)
+void OrthoCamera::SetPosition(const glm::dvec2& Position)
 {
 	this->Position = Position;
 }
 
 glm::mat4 OrthoCamera::GetViewMatrix() const
 {
-	return glm::lookAt(glm::vec3(this->Position, 0.1f), glm::vec3(this->Position, 0.0f), 
-		glm::vec3(0.0f, 1.0f, 0.0f));
+	return glm::lookAt(glm::dvec3(this->Position, 0.1), glm::dvec3(this->Position, 0.0), 
+		glm::dvec3(0.0, 1.0, 0.0));
 }
 
 glm::mat4 OrthoCamera::GetProjectionMatrix() const
 {
-	return glm::ortho(0.0f, this->ViewSize.x, 0.0f, this->ViewSize.y, 0.0f, 100.0f);
+	return glm::ortho(0.0, this->ViewSize.x, 0.0, this->ViewSize.y, 0.0, 100.0);
 }
 
-const glm::vec2& OrthoCamera::GetPosition() const { return this->Position; }
+const glm::dvec2& OrthoCamera::GetPosition() const { return this->Position; }
 
-const glm::vec2& OrthoCamera::GetViewSize() const { return this->ViewSize; }
+const glm::dvec2& OrthoCamera::GetViewSize() const { return this->ViewSize; }
