+++ SquareRunX/SquareRunX/Source/Engine/Graphics/OrthogonalCamera.h
@@ -4,18 +4,18 @@
 class OrthoCamera
 {
 private:
-	glm::vec2 Position;
-	glm::vec2 ViewSize;
+	glm::dvec2 Position;
+	glm::dvec2 ViewSize;
 public:
 	OrthoCamera();
-	OrthoCamera(const glm::vec2& Pos, const glm::vec2& ViewSize);
+	OrthoCamera(const glm::dvec2& Pos, const glm::dvec2& ViewSize);
 	~OrthoCamera();
 
-	void SetPosition(const glm::vec2& Position);
+	void SetPosition(const glm::dvec2& Position);
 	
 	glm::mat4 GetViewMatrix() const;
 	glm::mat4 GetProjectionMatrix() const;
 public:
-	const glm::vec2& GetPosition() const;
-	const glm::vec2& GetViewSize() const;
+	const glm::dvec2& GetPosition() const;
+	const glm::dvec2& GetViewSize() const;
 };
